KUDU-1741: Keep MiniCluster::Restart consistent with ExternalMiniCluster::Restart
MiniCluster and ExternalMiniCluster utility classes follow opposite
semantics for restart. For eg, ExternalMiniCluster::Restart() expects
that all the nodes must be shutdown before we restart, whereas
MiniCluster::Restart() expects that all servers are up and running
so that it can internally execute shutdown and start in that order.

This patch changes the semantics of MiniCluster::Restart() to follow
that of ExternalMiniCluster::Restart() to keep them consistent and
the existing tests are modified to follow the new semantics of
MiniCluster::Restart().

Change-Id: Iad33b7c46bfca3f277ccbca7d0420272f06a6633
Reviewed-on: http://gerrit.cloudera.org:8080/5598
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>