thirdparty: Link curl against el6 workaround openssl
This patch enables HTTPS support for thirdparty libcurl as well as links
thirdparty libcurl against the el6 workaround openssl when applicable.

Part of this patch pulled out of Todd's https://gerrit.cloudera.org/5015

This patch restores the capability to use HTTPS from libcurl for tests
while also maintaining support for posting to the dist_test cluster from
CentOS 6.

We considered (and attempted) depending on the system libcurl instead of
including it in thirdparty, however the system libcurl links against the
system openssl libraries. On CentOS 6, we build and test against the
CentOS 6.4 version of openssl, which does not provide the symbols needed
by the system version of libcurl. So, we are required to provide, as
part of thirdparty, any library needing to be linked against openssl,
and ensure that we build it against the el6-workaround openssl version.

Change-Id: I19e6be946853a80dad0f496839212887e76db06f
Reviewed-on: http://gerrit.cloudera.org:8080/5603
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>