Doc review for 1.2
I made a pass through all of the documentation and made improvements as
necessary for 1.2:

* Specify that the start/stop instructions are deb/rpm specific
* Updated the contribution C++ style guide to stop talking about
  gscoped_ptr
* Removed limitation that Spark is only for Spark 1 (now we support
  Spark 2)
* Fixed some branding in index.adoc
* Some small tweaks and improvements to the glossary of concepts

The larger changes are in the Impala integration. I tried to update the
syntax and examples to match the Impala-Kudu release from 11/22/2016
(the most recent release at the time of this writing and at the expected
release date for Kudu 1.2). Unfortunately, the syntax and capabilities
have changed again between that release and the upcoming Impala 2.8
release. Rather than document the unreleased syntax, I went with the
soon-to-be-removed syntax.

Similarly, I left in all of the instructions about how to install the
Impala-Kudu fork, even though that fork has now been merged back into
the main Impala release train. When Impala 2.8 is available, we will
have to do another pass here and re-publish docs.

Change-Id: I64ad4925ba4402ef3733a0ab62da91f1049848fb
Reviewed-on: http://gerrit.cloudera.org:8080/5676
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Tested-by: Kudu Jenkins
(cherry picked from commit 2505094c698c5a74fafde5f1d1ee6cd489ce836b)
Reviewed-on: http://gerrit.cloudera.org:8080/5693
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Todd Lipcon <todd@apache.org>