docs: update quickstart to a new dataset
The old dataset we were using seems to have changed formats
and no longer has unique identifiers, etc.

The new quickstart uses the SF MTA dataset, which is a bit nicer because
it's (a) bigger, (b) more time-series-like, and (c) has an associated
real-time feed so we could expand this tutorial more later.

Change-Id: If27de6e1b3d1a6178e036cbe68fd6812d5f74a94
Reviewed-on: http://gerrit.cloudera.org:8080/5701
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Tested-by: Kudu Jenkins