KUDU-1835 (part 1). Move compression codec classes to util
This creates a new library util_compression which has the compression
codec implementations. Previously, these were in cfile, but we don't
want to add a dependency from consensus onto cfile (that's a messy
coupling).

The reason to add this as a new module instead of just directly to util/
is that it brings in more third-party dependencies, and we've identified
that we'd like to try to slim down 'util' over time to make it easier
for other projects (eg Impala) to consume it piecemeal.

The other change here is that I had to move the CompressionType enum
from common.proto into a new proto module -- again to avoid an unwanted
dependency from util onto common.

The existing compression-test was half related to testing CFile
compression, and half a direct test of the codec implementations. As
such, the codec test was moved to the new module, and the cfile-related
test was moved into cfile-test.cc. As I did so, I realized that much of
the test content was redundant, so removed the redundant cases.

Change-Id: Iffe09aba18fd829f5918aabe061ca8b7f9d494c0
Reviewed-on: http://gerrit.cloudera.org:8080/5735
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>