[master] store CA information in the system table
The certificate authority information (private key and corresponding
CA certificate) is generated upon first start of Kudu cluster and
persistently stored in the system table.

Added simple unit tests for the corresponding entries in the
catalog table.  Also, added integration tests to verify the operation
in case of multi-master clusters.

The next step would be adding a way to re-generate CA information.

Change-Id: I16fa077e39f5d75f682cca7220371acdef4f0630
Reviewed-on: http://gerrit.cloudera.org:8080/5793
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Alexey Serbin <aserbin@cloudera.com>