Use Traits to avoid duplicate code in SSL wrappers
cert_management.cc contained some duplicate code for each of the
wrappers. In addition, having to always remember the correct 'free'
function for each SSL type was a bit difficult.

This uses a "traits" pattern to try to clean up the code. It doesn't
save a ton of lines of code, but it replaces a lot of imperative code
with more "declarative" lists of types and functions.

Change-Id: If1c44857358fc82f2694dff021b9f2ac3641313a
Reviewed-on: http://gerrit.cloudera.org:8080/5767
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Dan Burkert <danburkert@apache.org>