KUDU-1713: add a client Partitioner API
This adds an API which a client can use to determine which partition a
row falls into without actually writing that row. This is intended to be
used by Impala when writing into Kudu: Impala can pre-shuffle and
pre-sort the data being written so as to present a more efficient write
pattern to the Kudu tablets.

The corresponding Impala patch to use this API is at:
https://gerrit.cloudera.org/#/c/6559/

Change-Id: Ic08a078c75f15ef4200219b5260cfb99d79b72cc
Reviewed-on: http://gerrit.cloudera.org:8080/5775
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins