security: initial work on token creation and verification
This adds classes for managing Token Signing Keys (TSKs) on the
signer (masters) and verifiers (all servers). The new classes aren't
hooked up with the actual servers as of yet, nor do they actually use
real signature routines (blocked on another in-flight patch for that).

A unit test is included which drives the APIs using a stubbed-out
"signature" algorithm.

Change-Id: Iaf53ae50082d69028315952ac0732af6a83ffdbe
Reviewed-on: http://gerrit.cloudera.org:8080/5796
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins