KUDU-1853. Tablet copy: Don't orphan blocks on failure
Previously, if a tablet copy failed we would orphan data blocks. This
patch makes it so that a failed tablet copy operation that does not
involve a process crash does not orphan data blocks.

This also refactors some deletion logic out of TSTabletManager so that
TabletCopyClient will tombstone partially-copied tablets when the copy
operation fails.

Change-Id: Ifbc0720d460de8d912a241a5186ef6e8f524b830
Reviewed-on: http://gerrit.cloudera.org:8080/5799
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins