Fix flakiness of TestConnectionCache
This test would fail if one of the masters was a bit slow in starting. I
reproduced the failure reliably by adding a 3-second sleep in
ServerBase::Start.

This patch just makes a first dummy call that by design ensures that the
masters are available before we start pinging them.

Change-Id: Ib5061c002132901d9d1aab50400a7bdac4bc87d3
Reviewed-on: http://gerrit.cloudera.org:8080/5803
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>