Reduce default RSA key length for tests
The default RSA key length for real servers is set to 2048 bits.
However, generating a 2048-bit RSA key is relatively expensive. Since we
do it on every server start, and our tests start thousands of servers,
this has a negative impact on test time.

This changes the default in ExternalMiniCluster and in our own tests to
512-bit keys, which appear to be about 16x faster to generate compared
to 2048-bit keys (based on 'perf stat -r50 openssl genrsa -out /tmp/r
<key length>').

I measured the improvement on registration-test from ~4000ms to ~1200ms
wall time on my laptop.

This also changes the way we override flags for tests to use the gflags
API instead of directly setting FLAGS_* variables. This cleans up some
workarounds we did previously and also allows us to set flags which
might not apply for all test cases (eg security flags aren't in util
tests).

Change-Id: I310ac762aff4fed891a0491b2c3249e0faa9375f
Reviewed-on: http://gerrit.cloudera.org:8080/5804
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>