Upgrade Snappy to 1.1.4
In my random perusal of github, I noticed that we are a few point
releases out of date on Snappy. The release notes indicate that
compression should be ~5% faster and decompression 20% faster in 1.1.4.
v1.1.1 also claims to have increased decompression by 13-20% in some
particular benchmark cases.

Seems worth upgrading given the one-character diff.

Change-Id: Ieab52cc0c968f17b6a44328d49a42872284968bb
Reviewed-on: http://gerrit.cloudera.org:8080/5833
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Mike Percy <mpercy@apache.org>