KUDU-1854 Fixed broken links to Cloudera's Kudu documentation and archive.cloudera
With the 1.2 release, Cloudera's Kudu documentation was moved
out of the /betas/ folder. This broke any existing links we had to
the Cloudera docs. I made a pass through the docs and fixed the links.

The Kudu packages on archive.cloudera were also moved out of the /betas
directory. Those links were updated as well.

Change-Id: Ic158ddd6a576e7a3680d0f69c326768b7b3e5049
Reviewed-on: http://gerrit.cloudera.org:8080/5837
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>