Move 'master_rpc.{h,cc}' into the client/ module
This file has a couple of classes which are in fact only used by the
client and by the ExternalMiniCluster. Previously they'd been used by
the tablet server as well, so we ended up implementing them not in
'client/' but in 'master/'.

Now that the tablet server doesn't use these classes (it instead
heartbeats to all masters in parallel), it doesn't make much sense
anymore for them to be in the master/ directory.

This patch moves it into client/ and puts it in the
kudu::client::internal namespace. Additionally, one of the two classes
here is only used internally by the other, so I moved that into an
anonymous namespace in the .cc file.

This is in preparation for some surgery on this file to convert it into
more of a "connect to cluster" type RPC which will also be responsible
for grabbing an authentication token, the master CA list, etc.

Change-Id: Ic3a1db6aeb5326e6dfb9986b08949fa349b5f74d
Reviewed-on: http://gerrit.cloudera.org:8080/5864
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>