Allow setup-kudu-demo script to use a pre-built VM image
setup-kudu-demo-vm.sh script currently either downloads
a VM image or let us use an already downloaded image.
It is easier for future testing if we allow it to use a pre-built
VM image in situations where we want to use an image built out
of quickstart-vm repository and then run few kudu-examples.

This change lets the script take an OVA file as an argument
and retains the original usage of either downloading the image
or using a downloaded image if it is present in current directory.

Testing: Verified that all 3 modes of input work as expected.