[security] adjust TLS certificate verification
This commit adjusts when TLS certificate verification is performed in
the client and server, and adds a test of negotiating GSSAPI with TLS.
In the client, the server's cert is not verified when using the GSSAPI
SASL mech, since Kerberos provides strong authn. In the server, client
cert verification is disabled entirely. We'll have to turn it on in the
future selectively when we add support for client certs.

As part of this change, we no longer allow the SASL library to choose
the mechanism to use.  Instead, we determine the mechanism manually
using a simple heuristic: prefer GSSAPI to PLAIN. When we add support
for more mechanisms (e.g. KUDU_TOKEN), we'll update the heuristic
accordingly.

Change-Id: Id3b1698ccd8434b8d40d567e9d0fa506e4cdc0ca
Reviewed-on: http://gerrit.cloudera.org:8080/5865
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>