[mini_cluster] fix crash in GetLeaderMasterIndex()
If at least one of the mini-masters is shutdown by explicit call
of the MiniMaster::Shutdown() method, the subsequent call of
MiniCluster::GetLeaderMasterIndex() crashes while trying to access
the underlying Master object.  That's because the shared_ptr wrapper
has been reset by prior call of the MiniMaster::Shutdown() call.

This patch adds a check for the presence of the Master object in the
wrapper prior to call of Master::Shutdown().

Besides, added a way to specify no placeholder for the
index parameter for MiniCluster::GetLeaderMasterIndex().  That's
for the use case when it's necessary to wait for the master
to be established in the mini-cluster with no intent to get its index.

Change-Id: Icbb64ee44e22b5f32a2351e0e1289826a8485c84
Reviewed-on: http://gerrit.cloudera.org:8080/5870
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>