master: issue authentication tokens and CA certs to clients
This adds the code to the ConnectToMaster RPC to issue an authentication
token for the connecting user, as well as to send it the current master
CA cert.

The client side currently ignores both things, but a new unit test
verifies that they are getting properly set.

As with most of these early integration patches, there are plenty of
TODOs about integrating with the catalog manager, etc, but this should
be enough to unblock other client-side work of propagating the CA and
tokens into the RPC system.

Change-Id: I5969b8e125633b3b14364b98c0d0a992b162f302
Reviewed-on: http://gerrit.cloudera.org:8080/5871
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>