KUDU-1866: Add request-side sidecars
This patch adds sidecars to client requests. Using the same mechanism as
on the response-side, clients may attach slices to outbound requests
which do not pass through a serialization or copy before being pushed to
the network socket. On the server side, these sidecars may be read
directly from the underlying byte stream with the interposition of a
Protobuf wrapper.

The sidecars may be added to a request via RpcController and retrieved
via RpcContext (i.e. the reverse of the existing response-side
interface).

This patch adds a few tests to rpc-test, and all rpc-test tests pass.

Change-Id: I3d709edb2a22dc983f51b69d7660a39e8d8d6a09
Reviewed-on: http://gerrit.cloudera.org:8080/5908
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>