client: trust master's cert, adopt authn token
This changes the client to actually act on the security-related data
coming back from the ConnectToMaster() RPC. The client now adds the
master's CA cert to its trusted list, and adopts the authentication
token in a class member.

The token isn't used yet -- that's waiting on some concurrent work by
Dan to use tokens during RPC negotiation. But, a unit test verifies that
the token is getting set.

Change-Id: Iaa3eec178110b77d81608cfdce989e0fe494ee05
Reviewed-on: http://gerrit.cloudera.org:8080/5899
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>