Make SecureRpcHelper a Netty pipeline stage
This decouples SecureRpcHelper from TabletClient. When negotiation is
complete, it sends an "event" upstream to the TabletClient and removes
itself from the pipeline.

I'm hoping this decoupling will make it easier to test SecureRpcHelper
in isolation as we add more functionality like TLS negotiation.

After this is committed, I'm also hoping to rename SecureRpcHelper to
Negotiator.

Change-Id: I5f4b4516219b8eebf24786b4ceb13f2e6260f03b
Reviewed-on: http://gerrit.cloudera.org:8080/5927
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>