[security] tailored TokenSigner for system catalog
Updated the TokenSigner class in preparation for loading/storing TSKs
(Token Signing Keys) in system catalog.

The expected use-case for the TokenSigner is calling ImportKey() on
elected-as-leader callback with entries loaded from the system catalog
table, and then calling CheckNeedKey()/AddKey(), TryRotateKey() sequence.
Further down the road, it's necessary to call the CheckNeedKey()/AddKey(),
TryRotateKey() sequence periodically to generate and activate new TSKs.

Change-Id: Ie2417e2ccba6a1114db366b2f642f95362bf479c
Reviewed-on: http://gerrit.cloudera.org:8080/5930
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Dan Burkert <danburkert@apache.org>