java: use an outbound Encoder, add a test for Negotiator
This adds an Encoder to the front of the Netty pipeline, so that we can
send 'RpcOutboundMessage' objects instead of raw ChannelBuffers. This
doesn't really have any major effect, except that it makes testing a bit
easier -- we can stick parts of the pipeline in an Embedder and more
easily inspect the messages they are sending.

As proof of this, this patch adds a basic test for the Negotiator.

Change-Id: I75419b8c38dea9226ccac6a49ea00e17c74038f6
Reviewed-on: http://gerrit.cloudera.org:8080/5947
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Tested-by: Kudu Jenkins