java: add TLS support
This adds support for negotiating and speaking TLS in the Java client.

Testing is a bit light here, because the code on the C++ side to
generate and sign self-signed certs is still in flight, but once it's
committed, all of our integration tests should cover this code. I also
did write one unit test for the TLS handshake in particular.

I did test this against a local cluster by running kudu-ts against a
tserver and master compiled with the latest in-flight C++ patches to
generate self-signed certs. I verified with tshark that the traffic is
encrypted whereas I could read it before.

Change-Id: I9b095ccac908f391e683ec4ef433bae52d659a21
Reviewed-on: http://gerrit.cloudera.org:8080/5949
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>