java: allow eclipse checkstyle to work
Eclipse's checkstyle plugin doesn't properly handle paths relative to
the checkstyle configuration XML file. The workaround is that it does
provide a "samedir" property which can be used to construct the proper
absolute path. Of course, Maven doesn't provide that property by
default, so this patch also edits the pom so that Maven passes in "./"
as "samedir".

Change-Id: Ibb8c2d180b8c931b4e83ec5244f4b2e54155c5e4
Reviewed-on: http://gerrit.cloudera.org:8080/5943
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins